dashboard/api/requirements.txt
tensorflow==2.4.1      
tensorflow-hub==0.12.0      
flask==1.1.2 
pillow==7.2.0  
numpy==1.19.1
opencv-python==4.5.2.52